cdk/pyproject.toml
[tool.poetry]
name = "corporate-webhooks"
version = "0.1.0"
description = "CDK app for the Local Office Search API project"
authors = ["Citizens Advice"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
ca-cdk-constructs = {git = "https://github.com/citizensadvice/ca-cdk-constructs", rev = "v0.9.0"}
ruff = "^0.6.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.ruff]
target-version = 'py312'
line-length = 95

[tool.ruff.lint]
select = ['E', 'W', 'F', 'C', 'N', 'INP']
ignore = ['E501']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
